requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
openai==1.3.0
httpx==0.24.1
clickhouse-driver==0.2.6
pandas==2.1.3
numpy==1.26.2
matplotlib==3.8.2
langchain==0.0.350
langchain-community==0.0.10
faiss-cpu==1.7.4
sentence-transformers==2.2.2
huggingface-hub==0.16.4
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6
requests==2.31.0

